+++ SentryTestUtils/Sources/TestOptions.swift
@@ -9,9 +9,7 @@ public extension Options {
         enableAutoPerformanceTracing = false
         enableCrashHandler = false
         swiftAsyncStacktraces = false
-        #if !SDK_V10
-        enableAppHangTracking = false
-        #endif // !SDK_V10
+        appHangTimeoutInterval = 0
         enableNetworkTracking = false
         enableNetworkBreadcrumbs = false
         enableCaptureFailedRequests = false